Move `sanity` to under `phase_0`
