+++ tests/core/pyspec/eth2spec/test/phase_0/sanity/test_slots.py
